Reorganize quickPickWizard files
