+++ utils/src/index.ts
@@ -37,9 +37,9 @@ export * from './utils/contextUtils';
 export * from './activityLog/activities/ExecuteActivity';
 export * from './getAzExtResourceType';
 export * from './AzExtResourceType';
-export * from './treev2/quickPickWizard/experiences/appResourceExperience';
-export * from './treev2/quickPickWizard/experiences/compatibility/PickTreeItemWithCompatibility';
-export * from './treev2/quickPickWizard/experiences/contextValueExperience';
+export * from './pickTreeItem/experiences/appResourceExperience';
+export * from './pickTreeItem/experiences/compatibility/PickTreeItemWithCompatibility';
+export * from './pickTreeItem/experiences/contextValueExperience';
 export * from './utils/apiUtils';
 export * from './tree/isAzExtTreeItem';
 // NOTE: The auto-fix action "source.organizeImports" does weird things with this file, but there doesn't seem to be a way to disable it on a per-file basis so we'll just let it happen
