+++ utils/src/pickTreeItem/GenericQuickPickStep.ts
@@ -3,12 +3,12 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import * as types from '../../../index';
+import * as types from '../../index';
 import * as vscode from 'vscode';
-import { getLastNode } from './common/getLastNode';
-import { AzureWizardPromptStep } from '../../wizard/AzureWizardPromptStep';
-import { PickFilter } from './common/PickFilter';
-import { localize } from '../../localize';
+import { getLastNode } from './getLastNode';
+import { AzureWizardPromptStep } from '../wizard/AzureWizardPromptStep';
+import { PickFilter } from './PickFilter';
+import { localize } from '../localize';
 
 export interface GenericQuickPickOptions {
     skipIfOne?: boolean;
