+++ utils/src/pickTreeItem/contextValue/ContextValueQuickPickStep.ts
@@ -6,8 +6,8 @@
 import { TreeItem } from 'vscode';
 import * as types from '../../../index';
 import { parseContextValue } from '../../utils/contextUtils';
-import { PickFilter } from './common/PickFilter';
-import { GenericQuickPickOptions, GenericQuickPickStep } from './GenericQuickPickStep';
+import { PickFilter } from '../PickFilter';
+import { GenericQuickPickOptions, GenericQuickPickStep } from '../GenericQuickPickStep';
 
 export interface ContextValueFilterQuickPickOptions extends GenericQuickPickOptions {
     contextValueFilter: types.ContextValueFilter;
