+++ utils/src/pickTreeItem/contextValue/RecursiveQuickPickStep.ts
@@ -5,7 +5,7 @@
 
 import * as types from '../../../index';
 import { ContextValueFilterQuickPickOptions, ContextValueQuickPickStep } from './ContextValueQuickPickStep';
-import { getLastNode } from './common/getLastNode';
+import { getLastNode } from '../getLastNode';
 
 export class RecursiveQuickPickStep<TContext extends types.QuickPickWizardContext> extends ContextValueQuickPickStep<TContext, ContextValueFilterQuickPickOptions> {
     hideStepCount: boolean = true;
