+++ utils/src/pickTreeItem/experiences/compatibility/PickTreeItemWithCompatibility.ts
@@ -1,10 +1,10 @@
-import * as types from '../../../../../index';
+import * as types from '../../../../index';
 import * as vscode from 'vscode';
 import { ResourceGroupsItem } from '../../quickPickAzureResource/tempTypes';
 import { appResourceExperience } from '../appResourceExperience';
 import { subscriptionExperience } from '../subscriptionExperience';
-import { isAzExtTreeItem } from '../../../../tree/isAzExtTreeItem';
-import { createSubscriptionContext } from '../../../../utils/credentialUtils';
+import { isAzExtTreeItem } from '../../../tree/isAzExtTreeItem';
+import { createSubscriptionContext } from '../../../utils/credentialUtils';
 import { ISubscriptionContext } from '@microsoft/vscode-azext-dev';
 
 export namespace PickTreeItemWithCompatibility {
