+++ utils/src/pickTreeItem/quickPickAzureResource/QuickPickAppResourceStep.ts
@@ -4,10 +4,10 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { TreeItem } from 'vscode';
-import { AzureResourceQuickPickWizardContext } from '../../../../hostapi.v2';
-import * as types from '../../../../index';
-import { parseContextValue } from '../../../utils/contextUtils';
-import { PickFilter } from '../common/PickFilter';
+import { AzureResourceQuickPickWizardContext } from '../../../hostapi.v2';
+import * as types from '../../../index';
+import { parseContextValue } from '../../utils/contextUtils';
+import { PickFilter } from '../PickFilter';
 import { GenericQuickPickOptions, GenericQuickPickStep } from '../GenericQuickPickStep';
 import { AzureResourceItem } from './tempTypes';
